# azureTest/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "azureTest, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/UsdaInfo": {
      "get": {
        "tags": [
          "UsdaInfo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Metric",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Commodity",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Year",
            "type": "string"
          },
          {
            "in": "query",
            "name": "short_desc",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/GetUsdaDataRefactored": {
      "get": {
        "tags": [
          "UsdaInfo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Metric",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Commodity",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Year",
            "type": "string"
          },
          {
            "in": "query",
            "name": "short_desc",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/GetUsdaDataRefactoredMultiYear": {
      "get": {
        "tags": [
          "UsdaInfo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Metric",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Commodity",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Year",
            "type": "string"
          },
          {
            "in": "query",
            "name": "short_desc",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/GetByState": {
      "get": {
        "tags": [
          "UsdaInfo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Metric",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Commodity",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Year",
            "type": "string"
          },
          {
            "in": "query",
            "name": "short_desc",
            "type": "string"
          },
          {
            "in": "query",
            "name": "stateAlpha",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UsdaInfoOld": {
      "get": {
        "tags": [
          "UsdaInfo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Metric",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Commodity",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Year",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/getnoparams": {
      "get": {
        "tags": [
          "UsdaInfo"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  }
}